# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification for مقرأة سَنا الآي Quran Center Management System
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: Invalid credentials
  /auth/login/select:
    post:
      operationId: loginSelectAccount
      tags: [auth]
      summary: Select account after multi-role login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginSelectBody"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: Invalid credentials
...
  /auth/me:
    get:
      operationId: getCurrentUser
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
...
  /badge-events:
    get:
      operationId: listBadgeEvents
      tags: [badges]
      responses:
        "200":
          description: Badge events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BadgeEvent"
    post:
      operationId: createBadgeEvent
      tags: [badges]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBadgeEventBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BadgeEvent"
  /badge-events/{id}/auto-assign:
    post:
      operationId: autoAssignBadgeEvent
      tags: [badges]
      summary: Auto-assign badge to all qualifying entities
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Auto-assigned result
          content:
            application/json:
              schema:
                type: object
                properties:
                  assigned:
                    type: integer
                  skipped:
                    type: integer
                  message:
                    type: string
                required: [assigned, skipped, message]
  /badge-events/{id}:
    patch:
      operationId: updateBadgeEvent
      tags: [badges]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBadgeEventBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BadgeEvent"
    delete:
      operationId: deleteBadgeEvent
      tags: [badges]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /badge-assignments:
    get:
      operationId: listBadgeAssignments
      tags: [badges]
      parameters:
        - name: entityType
          in: query
          schema:
            type: string
        - name: entityId
          in: query
          schema:
            type: integer
        - name: badgeEventId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Badge assignments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BadgeAssignment"
    post:
      operationId: createBadgeAssignment
      tags: [badges]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBadgeAssignmentBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BadgeAssignment"
  /badge-assignments/{id}:
    delete:
      operationId: deleteBadgeAssignment
      tags: [badges]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    LoginBody:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required:
        - email
        - password
    LoginSelectBody:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
        accountId:
          type: integer
      required:
        - email
        - password
        - accountId
    LoginResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
        requiresSelection:
          type: boolean
        accounts:
          type: array
          items:
            $ref: "#/components/schemas/AccountOption"
...
    User:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [leader, data_entry, teacher, supervisor, student, track_supervisor, volunteer, exam_supervisor]
        track:
          type: ["string", "null"]
        circleId:
          type: ["integer", "null"]
        phone:
          type: ["string", "null"]
        country:
          type: ["string", "null"]
        ageRange:
          type: ["string", "null"]
        educationLevel:
          type: ["string", "null"]
        isArchived:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - email
        - name
        - role
        - isArchived
        - createdAt
...
    AccountOption:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        role:
          type: string
        track:
          type: ["string", "null"]
        circleId:
          type: ["integer", "null"]
      required:
        - id
        - name
        - role
...
    BadgeEvent:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        emoji:
          type: string
        color:
          type: string
        targetType:
          type: string
        dateFrom:
          type: string
        dateTo:
          type: string
        isActive:
          type: boolean
        createdById:
          type: integer
        createdAt:
          type: string
        assignmentCount:
          type: integer
      required: [id, name, emoji, color, targetType, dateFrom, dateTo, isActive, createdById, createdAt, assignmentCount]
    CreateBadgeEventBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]
        emoji:
          type: string
        color:
          type: string
        targetType:
          type: string
        dateFrom:
          type: string
        dateTo:
          type: string
        isActive:
          type: boolean
      required: [name, emoji, color, targetType, dateFrom, dateTo]
    BadgeAssignment:
      type: object
      properties:
        id:
          type: integer
        badgeEventId:
          type: integer
        badgeName:
          type: string
        badgeEmoji:
          type: string
        badgeColor:
          type: string
        entityType:
          type: string
        entityId:
          type: integer
        entityName:
          type: string
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, badgeEventId, badgeName, badgeEmoji, badgeColor, entityType, entityId, entityName, createdAt]
    CreateBadgeAssignmentBody:
      type: object
      properties:
        badgeEventId:
          type: integer
        entityType:
          type: string
        entityId:
          type: integer
        entityName:
          type: string
        notes:
          type: ["string", "null"]
      required: [badgeEventId, entityType, entityId, entityName]
...